requirements.txt
pandas
numpy
matplotlib
openpyxl
tabulate
streamlit
fpdf2